@@ -9,6 +9,7 @@ namespace ble
 class CGattService
 {
 public:
+	CGattService() = default;
 	CGattService(uint16_t uuid, std::vector<CCharacteristic>& characteristics);
 	CGattService(const CGattService& other);
 	CGattService(CGattService&& other) = default;
@@ -16,7 +17,6 @@ public:
 	CGattService& operator=(CGattService&& other) = default;
 	explicit operator ble_gatt_svc_def() const;
 private:
-	CGattService() = default;
 	[[nodiscard]] CGattService copy(const CGattService& source) const;
 private:
 	std::unique_ptr<ble_uuid128_t> m_pUUID;
